27 min into the ticket video
