+++ src/services/employeeServices.jsx
@@ -1,0 +1,7 @@
+// src/services/employeeServices.jsx
+export const getAllEmployees = () => {
+    return fetch("http://localhost:8088/employees?_expand=user").then((res) =>
+      res.json()
+    );
+  };
+
