+++ src/services/ticketService.jsx
@@ -1,4 +1,5 @@
+// src/components/ticketServices.jsx
 export const getAllTickets = () => {
     
-    return fetch(`http://localhost:8088/serviceTickets`).then(res => res.json());
+    return fetch(`http://localhost:8088/serviceTickets?_embed=employeeTickets`).then(res => res.json());
 }
